* profil.h (PROFADDR): Prevent overflow when text segments are larger than 256k. * profil.c (profthr_func): Raise thread priority for more accurate sampling. * path.cc (hash_path_name): Use ino_t as type.

+++ winsup/cygwin/profil.h
@@ -24,7 +24,7 @@ details. */
 
 /* convert an index into an address */
 #define PROFADDR(idx, base, scale)	\
-	((base) + ((((idx) << 16) / (scale)) << 1))
+	((base) + ((((unsigned long long)(idx) << 16) / (scale)) << 1))
 
 /* convert a bin size into a scale */
 #define PROFSCALE(range, bins)		(((bins) << 16) / ((range) >> 1))
